// src/api/openapi-client/models/ConversationCreate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Retirement‐Matching API
 * REST endpoints for user auth (via magic‐link), retirees, career elements, reference data, startups, job postings, matching, conversations, and hire invites. Uses a JWT set in a cookie. 
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ConversationCreate
 */
export interface ConversationCreate {
    /**
     * 
     * @type {Set<string>}
     * @memberof ConversationCreate
     */
    participantIds: Set<string>;
    /**
     * 
     * @type {string}
     * @memberof ConversationCreate
     */
    jobPostingId?: string;
}

/**
 * Check if a given object implements the ConversationCreate interface.
 */
export function instanceOfConversationCreate(value: object): value is ConversationCreate {
    if (!('participantIds' in value) || value['participantIds'] === undefined) return false;
    return true;
}

export function ConversationCreateFromJSON(json: any): ConversationCreate {
    return ConversationCreateFromJSONTyped(json, false);
}

export function ConversationCreateFromJSONTyped(json: any, ignoreDiscriminator: boolean): ConversationCreate {
    if (json == null) {
        return json;
    }
    return {
        
        'participantIds': new Set(json['participantIds']),
        'jobPostingId': json['jobPostingId'] == null ? undefined : json['jobPostingId'],
    };
}

export function ConversationCreateToJSON(json: any): ConversationCreate {
    return ConversationCreateToJSONTyped(json, false);
}

export function ConversationCreateToJSONTyped(value?: ConversationCreate | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'participantIds': Array.from(value['participantIds'] as Set<any>),
        'jobPostingId': value['jobPostingId'],
    };
}