// src/api/openapi-client/models/ConversationSummary.ts
export interface ConversationSummary {
    /**
     * Unique identifier for this conversation.
     * @type {string}
     * @memberof ConversationSummary
     */
    id: string;
    /**
     * When the conversation was created.
     * @type {Date}
     * @memberof ConversationSummary
     */
    createdAt: Date;
    /**
     * 
     * @type {JobPosting}
     * @memberof ConversationSummary
     */
    jobPosting?: JobPosting;
    /**
     * Unique identifier for the participating retiree.
     * @type {Array<UserPublicSummary>}
     * @memberof ConversationSummary
     */
    participants: Array<UserPublicSummary>;
    /**
     * Number of unread messages in this conversation.
     * @type {number}
     * @memberof ConversationSummary
     */
    unreadMessageCount: number;
    /**
     * 
     * @type {Message}
     * @memberof ConversationSummary
     */
    lastMessage?: Message;
}