// src/api/openapi-client/models/MessageCreate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Retirement‐Matching API
 * REST endpoints for user auth (via magic‐link), retirees, career elements, reference data, startups, job postings, matching, conversations, and hire invites. Uses a JWT set in a cookie. 
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface MessageCreate
 */
export interface MessageCreate {
    /**
     * 
     * @type {string}
     * @memberof MessageCreate
     */
    content: string;
}

/**
 * Check if a given object implements the MessageCreate interface.
 */
export function instanceOfMessageCreate(value: object): value is MessageCreate {
    if (!('content' in value) || value['content'] === undefined) return false;
    return true;
}

export function MessageCreateFromJSON(json: any): MessageCreate {
    return MessageCreateFromJSONTyped(json, false);
}

export function MessageCreateFromJSONTyped(json: any, ignoreDiscriminator: boolean): MessageCreate {
    if (json == null) {
        return json;
    }
    return {
        
        'content': json['content'],
    };
}

export function MessageCreateToJSON(json: any): MessageCreate {
    return MessageCreateToJSONTyped(json, false);
}

export function MessageCreateToJSONTyped(value?: MessageCreate | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'content': value['content'],
    };
}